WIP. Progressing through various errors.
Cleaned up little things and one big thing (bare variable in with_items)
found by ansible-lint.

Currently stuck on ansible, systemd, and collectd not all speaking the
same language, causing a failure to start the service. I'm going to try
running in a vm instead of a container to see if that helps, so I wanted
a checkpoint commit first.